changes to work with llvm i.e. removal of try/catch
