+++ src/Compile.cc
@@ -1,0 +1,44 @@
+#include "AST.hh"
+#include <iostream>
+
+void Expression::generate_llvm_ir() {
+    // TODO: implement
+    std::cerr << "Not Implemented";
+    exit(1);
+}
+
+void UnaryOperator::generate_llvm_ir() {
+    // TODO: implement
+    std::cerr << "Not Implemented";
+    exit(1);
+}
+
+void BinaryOperator::generate_llvm_ir() {
+    // TODO: implement
+    std::cerr << "Not Implemented";
+    exit(1);
+}
+
+void IfOperator::generate_llvm_ir() {
+    // TODO: implement
+    std::cerr << "Not Implemented";
+    exit(1);
+}
+
+void FunctionCall::generate_llvm_ir() {
+    // TODO: implement
+    std::cerr << "Not Implemented";
+    exit(1);
+}
+
+void ValueDef::generate_llvm_ir() {
+    // TODO: implement
+    std::cerr << "Not Implemented";
+    exit(1);
+}
+
+void FunctionDef::generate_llvm_ir() {
+    // TODO: implement
+    std::cerr << "Not Implemented";
+    exit(1);
+}
